Recycler adapter: Reloading from DB also followed sessions
