+++ app/src/main/java/pl/llp/aircasting/screens/dashboard/following/FollowingRecyclerAdapter.kt
@@ -6,7 +6,6 @@ import androidx.fragment.app.FragmentManager
 import pl.llp.aircasting.models.Session
 import pl.llp.aircasting.screens.dashboard.SessionCardListener
 import pl.llp.aircasting.screens.dashboard.SessionsRecyclerAdapter
-import java.util.*
 
 open class FollowingRecyclerAdapter(
     private val mInflater: LayoutInflater,
@@ -26,8 +25,7 @@ open class FollowingRecyclerAdapter(
     }
 
     override fun prepareSession(session: Session, expanded: Boolean): Session {
-        // We only have to reload measurements for fixed tab for expanded sessions. Followed sessions have measurements fetched anyway
-        return session
+        return reloadSessionFromDB(session)
     }
 
 }
